code/host/host/pyproject.toml
[tool.poetry]
name = "host"
version = "0.1.0"
description = ""
authors = ["lheywang <leonard.heywang@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pycaw = "^20240210"
pulsectl = "^24.12.0"
pillow = "^11.2.1"
cairosvg = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
